phase2/requirements.txt
cerebras-cloud-sdk>=1.0.0
pydantic>=2.7
rich>=13.7
orjson>=3.10
python-dotenv>=1.0
